+++ src/components/AnimatedSection.js
