+++ src/components/Chatbot.js
@@ -2,11 +2,11 @@ import React, { useState, useEffect, useCallback, useMemo } from 'react';
 import {Flex, VStack, useToast, Box, Button} from '@chakra-ui/react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import api, { aiApi } from "../apis/api";
-import Sidebar from '../component/chatbot/ChSideBar';
-import Header from '../component/chatbot/Chheader';
-import ChatMessage from '../component/chatbot/ChatMessage';
-import ImageGallery from '../component/chatbot/ImageGallery';
-import InputArea from '../component/chatbot/InputArea';
+import Sidebar from './/chatbot/ChSideBar';
+import Header from './/chatbot/Chheader';
+import ChatMessage from './/chatbot/ChatMessage';
+import ImageGallery from './/chatbot/ImageGallery';
+import InputArea from './/chatbot/InputArea';
 import jsPDF from 'jspdf';
 import koreanFontUrl from '../font/NanumMyeongjo-Regular.ttf';
 import axios from "axios";
