+++ src/components/MarketGrowthChart.js
