+++ src/components/PrivacyText.js
