+++ src/components/chatbot/ImageGallery.js
